[Dataset] transforms fix bugs, debugging and add  group normalize method
